% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.R
\name{adjacted_move}
\alias{adjacted_move}
\title{Create adjacent movement matrix}
\usage{
adjacted_move(n_rows, n_cols, prob)
}
\arguments{
\item{n_rows}{number of rows in population matrix}

\item{n_cols}{number of columns in population matrix}

\item{prob}{probability of movement}
}
\description{
Create a transition matrix of movement to adjacent cells
}
