% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.r
\name{baranov}
\alias{baranov}
\title{Baranov catch equation}
\usage{
baranov(N, M, F)
}
\arguments{
\item{N}{population size}

\item{M}{natural mortality (instantaneous)}

\item{F}{fishing mortality (instantaneous)}
}
\description{
Returns catch for given F for population size N
with natural and fishing mortality M and F respectively
}
