% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvest.R
\name{calc_catch}
\alias{calc_catch}
\title{Calculate catch}
\usage{
calc_catch(control, N_area, est_area)
}
\arguments{
\item{control}{a control file}

\item{N_area}{true population numbers, can be by area}

\item{est_area}{estimated population numbers, can be by area}
}
\description{
Calculate catch numbers for a year
}

