% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{catch_to_F}
\alias{catch_to_F}
\title{Calculate F for a given catch}
\usage{
catch_to_F(N, nat_mort, fish_mort, catch)
}
\arguments{
\item{N}{population size}

\item{nat_mort}{instantaneous natural mortality}

\item{fish_mort}{instantaneous fishing mortality}

\item{catch}{catch}
}
\description{
Root of this function is the F that corresponds to a given catch
}
\details{
Based on code written by Bill de la Mare
}
