% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_hcr.R
\name{create_control_hcr}
\alias{create_control_hcr}
\alias{control}
\title{Create a simplified control file for a single-area population without tags}
\usage{
create_control_hcr(years, pop_pars, rec_pars = list(type = "constant", mu = 0,
  s = 1, spat_dist = "uniform"), harvest_pars, assess_pars, stoch_rec = TRUE)
}
\arguments{
\item{years}{vector of years (integers starting with 1)}

\item{pop_pars}{list of population parameters containing the initial population
size ("initial") and instantaneous natural mortality ("nat_mort")}

\item{rec_pars}{list of recruitment parameters containing the name of the mechanism
for calculating recruitment ("type"; can be either "constant", "logistic", "bevholt"
or "ricker") see \code{\link{est_recruits}}}

\item{harvest_pars}{harvest parameters a list containing "type" ("const_exploit" =
constant exploitation rate, "TAC" = constant catch, ...),  (see \code{\link{calc_catch}})}

\item{assess_pars}{list of assessment parameters (see \code{\link{do_assessment}})}

\item{stoch_rec}{stochasitc recruitment (default=TRUE)}
}
\description{
The control file specifies the simulation parameters
}
\details{
A simulation is run using the function (see \code{\link{run_sim}})
}
