% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{create_model}
\alias{create_model}
\title{Create a model object}
\usage{
create_model(control)
}
\arguments{
\item{control}{a control file}
}
\description{
Create a list of arrays of the appropriate dimensions for storing population
and fishery information
}
\details{
Note tags could be a multi-dimensional array to account for tag movement
}
