% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_hcr.R
\name{create_model_hcr}
\alias{create_model_hcr}
\title{Create a simplified model object for a single-area population without tags}
\usage{
create_model_hcr(control)
}
\arguments{
\item{control}{a control file}
}
\description{
Create a list of arrays of the appropriate dimensions for storing population
and fishery information
}
\details{
Note tags could be a multi-dimensional array to account for tag movement
}
