% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{create_model_multi_release}
\alias{create_model_multi_release}
\title{Create a multiple release model object}
\usage{
create_model_multi_release(control)
}
\arguments{
\item{control}{a control file}
}
\description{
Create a list of arrays of the appropriate dimensions for storing population
and fishery information
Note tags could be a multi-dimensional array to account for tag movement
}
