% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{create_storage}
\alias{create_storage}
\title{Create an object to store simulation results}
\usage{
create_storage(control, nsims)
}
\arguments{
\item{control}{a control file}

\item{nsims}{number of simulations}
}
\description{
Create an object to store simulation results
}
