% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_hcr.R
\name{create_storage_hcr}
\alias{create_storage_hcr}
\title{Create an object to store a hcr simulation results}
\usage{
create_storage_hcr(control, n_reps)
}
\arguments{
\item{control}{a control file}

\item{n_reps}{number of replicate simulations}
}
\description{
Create an object to storea hcr simulation results
}
