% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessment.R
\name{do_assessment}
\alias{do_assessment}
\title{Survey a population}
\usage{
do_assessment(control, model, year)
}
\arguments{
\item{control}{a control file}

\item{model}{a tempory file created by \code{\link{run_sim}}}

\item{year}{survey year}
}
\description{
Wrapper for functions to obtain estimates of population size
during a simulation
}
