% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruitment.R
\name{est_recruits}
\alias{est_recruits}
\title{Estimate recruitment}
\usage{
est_recruits(type, N_area, rec_pars)
}
\arguments{
\item{type}{either "constant", "logistic", "bevholt" or "ricker"}

\item{N_area}{population by area}

\item{rec_pars}{recruitment parameters (from control file)}
}
\description{
Estimate recruitment the total annual recruitment
}
