% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{F_to_catch}
\alias{F_to_catch}
\title{Calculates catch for a given F}
\usage{
F_to_catch(N, nat_mort, fish_mort)
}
\arguments{
\item{N}{population size}

\item{nat_mort}{instantaneous natural mortality}

\item{fish_mort}{instantaneous fishing mortality}
}
\description{
Returns catch from N for a given F, used in \code{catchtoFfinding} the F that
corresponds to a given catch.
}
\details{
Based on code written by Bill de la Mare
}
