% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvest.R
\name{fishing_locs}
\alias{fishing_locs}
\title{Select fishing locations}
\usage{
fishing_locs(type, regions, prop, revisit_prop, old_locs)
}
\arguments{
\item{type}{fishing strategy}

\item{regions}{vector of regions}

\item{prop}{proportion of cells to be fished}

\item{revisit_prop}{proportion of locations to revisit}

\item{old_locs}{vector of locations fished previous year}
}
\description{
Fishing locations based on a strategy and proportion area fished
}
