% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_parameters.R
\name{init_k}
\alias{init_k}
\title{Calculate the equilibrium population size}
\usage{
init_k(pop_pars, rec_pars, max_reps = 1000, tol = 1)
}
\arguments{
\item{pop_pars}{population parameters including the initial population size
"initial" and natural mortality "nat_mort"}

\item{rec_pars}{recruitment parameters}
}
\description{
Calculate the equilibrium population size. Can be used as an estimate of
k
}
