% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.R
\name{move_matrix}
\alias{move_matrix}
\title{Create movement matrix}
\usage{
move_matrix(method, n_rows, n_cols, prob = NULL, matrix = NULL)
}
\arguments{
\item{method}{movement method}

\item{n_rows}{number of rows in population matrix}

\item{n_cols}{number of columns in population matris}

\item{prob}{movement probability}

\item{matrix}{user supplied movement matrix}
}
\description{
Create a transition matrix of movement probabilities
}

