% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.R
\name{move_N}
\alias{move_N}
\title{Move a fish population}
\usage{
move_N(d, move_mat)
}
\arguments{
\item{d}{vector of starting population by region}

\item{move_mat}{movement transition matrix of}
}
\description{
Implement movement of fish in a population array
}
