% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_release.R
\name{multi_release}
\alias{multi_release}
\title{Multiple tag release estimate of population size}
\usage{
multi_release(tags, hauls, pars)
}
\arguments{
\item{tags}{matrix (or dataframe?) where the first column specifies the annual releases and
subsequent columns the annual recaptures by release cohort (rows). Note 
these must be whole numbers}

\item{hauls}{matrix (or dataframe?) where the first column represents the
untagged recaptures and subsequent columns the recaptures by release cohort 
with the first (oldest) cohort to the left and the last (youngest) to the
right}

\item{pars}{list of model parameters}
}
\description{
Estimate population size from multiple release, single recapture tagging data
}
\details{
For a Type 1 fishery the adjusted releases are calculate as follows

1) Initial tag-induced mortality is applied

2) Within season recaptures divided by the tag reporting rate are removed

3) Natural mortality and chronic tag loss and tag-induced mortality is applied.

For a fishery that operates throughout the year (Ricker Type 2) fishing and
natural mortality competed to deplete the tagged population. The adjusted
releases are calculated as follows

1) Initial tag-induced mortality is applied

2) Half of the within season recaptures divided by the tag reporting rate are
removed

3) Half of the natural mortality and chronic tag loss and tag-induced
mortality is applied

4) The remaining within season recaptures divided by the tag reporting rate 
are removed

5) The remaining natural mortality and chronic tag loss and tag-induced
 mortality is applied.
}
\seealso{
\code{\link{pars}}
}
