% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{petersen}
\alias{petersen}
\alias{chapman_n}
\alias{chapman_wt}
\alias{petersen}
\title{Petersen estimators}
\usage{
chapman_n(tags, catch, recaps, check_type = "srelease")

chapman_wt(tags, catch, recaps, mean_wt = 0, check_type = "srelease")

petersen(tags, catch, recaps, check_type = "srelease")
}
\arguments{
\item{tags}{number of marked animals released (note these must be whole 
numbers for the bootstrap to work)}

\item{catch}{vector of the number or weight of animals captured and checked 
for tags by haul on the second survey}

\item{recaps}{vector of number of marked animals recaptured by haul}

\item{check_type}{of input checks "srelease" (default= single release checks)
other options are "mrelease" (multiple release checks) and "zero_recaps"
(include}

\item{mean_wt}{mean weight of a single fish (optional argument for Chapman
weight method)}
}
\description{
Single tag recapture estimates Petersen, Lincoln Petersen estimators
}
