% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_release.R
\name{print.srelease}
\alias{print.srelease}
\title{Print method for single release, single recapture models}
\usage{
\method{print}{srelease}(x, ...)
}
\arguments{
\item{x}{object of class srelease}

\item{...}{additional parameters}
}
\description{
S3 method for single release, single recapture models 
tag recapture models of abundance
}
