% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruitment.R
\name{recruit_area}
\alias{recruit_area}
\title{Assign recruitment proportions by area}
\usage{
recruit_area(method, n_areas, user_props = NULL, ...)
}
\arguments{
\item{method}{current options include "uniform", "lognormal" and "user" defined}

\item{n_areas}{number of model areas}

\item{user_props}{user defined proportions}

\item{...}{additional parameters}
}
\description{
Calculate the proportions of recruitment by spatial cell
}
