% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_hcr.R
\name{run_hcr}
\alias{run_hcr}
\title{Run a single-area simulation with a harvest control rule}
\usage{
run_hcr(control, n_reps, save = FALSE, path = NULL, filename = NULL)
}
\arguments{
\item{control}{control file}

\item{n_reps}{number of replicate simulations}

\item{save}{save the output (default=FALSE)}

\item{path}{path for the output}

\item{filename}{filename for the control file}
}
\description{
Run a simulation using a harvest control rule to modify catch during the
simulation
}
\examples{
## add an example
}
