% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sim.R
\name{run_sim}
\alias{run_sim}
\title{Run a simulation model}
\usage{
run_sim(control, n_reps)
}
\arguments{
\item{control}{control file}

\item{n_reps}{number of MCMC replicates}
}
\description{
Run a simulation model
}
