% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sim_multi_release_process_reorderd.R
\name{run_sim_multi_release_new_process_order}
\alias{run_sim_multi_release_new_process_order}
\title{Run a simulation model}
\usage{
run_sim_multi_release_new_process_order(control, n_reps, run_assessment)
}
\arguments{
\item{control}{control file}

\item{n_reps}{number of MCMC replicates}

\item{run_assessment}{TRUE if you want the simulation to estimate biomass and FALSE if estimates are not required}
}
\description{
Run a simulation model
}
