% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_release.R
\name{summary.bsamples}
\alias{summary.bsamples}
\title{S3 method for bootstrapped confidence intervals}
\usage{
\method{summary}{bsamples}(object, quantiles = c(0.025, 0.5, 0.975), ...)
}
\arguments{
\item{object}{object of class bsamples}

\item{quantiles}{bootstrap quantiles (default 0.025, 0.5, 0.975)}

\item{...}{additional parameters}
}
\description{
Calculate bootstrapped confidence intervals for object of
class bsamples
}
