% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_release.R
\name{summary.srelease}
\alias{summary.srelease}
\title{Summary method for single release, single recapture models}
\usage{
\method{summary}{srelease}(object, ...)
}
\arguments{
\item{object}{object of class srelease}

\item{...}{additional parameters}
}
\description{
S3 method for single release, single recapture models 
tag recapture models of abundance
}
