% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.r, R/recruitment.R
\name{est_recruits}
\alias{est_recruits}
\title{Calculate recrutiment from previous years population}
\usage{
est_recruits(type, N_area, rec_pars, var = FALSE)

est_recruits(type, N_area, rec_pars, var = FALSE)
}
\arguments{
\item{type}{recruitment function either "constant", "logistic", "bevholt" (Beverton Holt)
or "ricker". The "lognorm" method is independent of population size}

\item{N_area}{population size by region}

\item{rec_pars}{list of recruitment parameters}

\item{var}{is recruitment stochastic (default=FALSE)}

\item{type}{either "constant", "logistic", "bevholt" or "ricker"}

\item{N_area}{population by area}

\item{rec_pars}{recruitment parameters (from control file)}

\item{var}{recruitment variability (default FALSE)}
}
\description{
Function returns number of recruits given last year's population size
carrying capacity K, natural survivorship S and resilience r.
rk is the number of recruits per unit spawner when B = K.
var controls whether the recruitment is stochastic
(var = F is used when calculating recruitment parameters)

Estimate recruitment the total annual recruitment
}

