% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruitment.R
\name{est_recruits}
\alias{est_recruits}
\title{Calculate recrutiment from previous years population}
\usage{
est_recruits(type, N_area, rec_pars)
}
\arguments{
\item{type}{recruitment function either "constant", "logistic", "bevholt"
(Beverton Holt) or "ricker". The "lognorm" method is independent of
population size}

\item{N_area}{population size by region}

\item{rec_pars}{list of recruitment parameters (note: extend this help)}
}
\description{
Function returns number of recruits given last year's population size N,
carrying capacity k and resilience r. rk is the number of recruits per unit
spawner when B = k.
}
\details{
Modification of function written by Bill de la Mare
}
