% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sim.R
\name{run_sim}
\alias{run_sim}
\title{Run a simulation model}
\usage{
run_sim(control, n_reps, save = FALSE, path = NULL, filename = NULL)
}
\arguments{
\item{control}{control file}

\item{n_reps}{number of replicate simulations}

\item{save}{save the output (default=FALSE)}

\item{path}{path for the output}

\item{filename}{filename for the control file}
}
\description{
Run a simulation model
}
