% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sim.R
\name{run_sim}
\alias{run_sim}
\title{Run a simulation model}
\usage{
run_sim(control, n_reps, save = FALSE, path = NULL, f_name = NULL)
}
\arguments{
\item{control}{control file}

\item{n_reps}{number of MCMC replicates}

\item{save}{save the output (default=FALSE)}

\item{path}{path for the output}

\item{f_name}{filename for the control file (or should this be the file name
for the output?)}
}
\description{
Run a simulation model
}
